Fix foreach narrow index on -m64.

+++ compiler_tests/test77.d
@@ -1,11 +1,11 @@
 //T compiles:yes
 //T retval:0
-//T has-passed:no
+//T has-passed:yes
 
 void main()
 {
 	string str = "foobar";
 
     // This is narrowing, but valid.
-	foreach(int i, c; str) {}
+	foreach(byte i, c; str) {}
 }
